Move sample to playground
