--- TestProjects/Logging.Core/App.cs
@@ -1,17 +1,0 @@
-using MvvmCross.Platform.IoC;
-
-namespace Logging.Core
-{
-    public class App : MvvmCross.Core.ViewModels.MvxApplication
-    {
-        public override void Initialize()
-        {
-            CreatableTypes()
-                .EndingWith("Service")
-                .AsInterfaces()
-                .RegisterAsLazySingleton();
-
-            RegisterAppStart<ViewModels.MainViewModel>();
-        }
-    }
-}
