--- TestProjects/Logging.Core/Logging.Droid/Assets/AboutAssets.txt
@@ -1,19 +1,0 @@
-Any raw assets you want to be deployed with your application can be placed in
-this directory (and child directories) and given a Build Action of "AndroidAsset".
-
-These files will be deployed with your package and will be accessible using Android's
-AssetManager, like this:
-
-public class ReadAsset : Activity
-{
-	protected override void OnCreate (Bundle bundle)
-	{
-		base.OnCreate (bundle);
-
-		InputStream input = Assets.Open ("my_asset.txt");
-	}
-}
-
-Additionally, some Android functions will automatically load asset files:
-
-Typeface tf = Typeface.CreateFromAsset (Context.Assets, "fonts/samplefont.ttf");
