--- TestProjects/Logging.Core/Logging.Droid/Properties/AndroidManifest.xml
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<manifest xmlns:android="http://schemas.android.com/apk/res/android" android:versionCode="1" android:versionName="1.0" package="com.willsb.Logging">
-	<uses-sdk android:minSdkVersion="10" />
-	<application android:allowBackup="true" android:icon="@mipmap/icon" android:label="@string/app_name">
-	</application>
-</manifest>
