--- TestProjects/Logging.Core/Logging.Droid/Resources/values/Strings.xml
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<resources>
-	<string name="hello">Hello World, Click Me!</string>
-	<string name="app_name">Logging</string>
-</resources>
