--- TestProjects/Logging.Core/Logging.Droid/Views/MainView.cs
@@ -1,16 +1,0 @@
-using Android.App;
-using Android.OS;
-using MvvmCross.Droid.Views;
-
-namespace Logging.Droid.Views
-{
-    [Activity(Label = "View for MainViewModel")]
-    public class MainView : MvxActivity
-    {
-        protected override void OnCreate(Bundle bundle)
-        {
-            base.OnCreate(bundle);
-            SetContentView(Resource.Layout.Main);
-        }
-    }
-}
