--- TestProjects/Logging.Core/Logging.Droid/packages.config
@@ -1,30 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.CSharp" version="4.0.1" targetFramework="monoandroid80" />
-  <package id="Serilog" version="2.6.0-dev-00915" targetFramework="monoandroid80" />
-  <package id="Serilog.Sinks.Console" version="3.0.1" targetFramework="monoandroid80" />
-  <package id="Serilog.Sinks.Literate" version="3.0.1-dev-00044" targetFramework="monoandroid80" />
-  <package id="Serilog.Sinks.Xamarin" version="0.1.29" targetFramework="monoandroid80" />
-  <package id="System.Collections" version="4.0.11" targetFramework="monoandroid80" />
-  <package id="System.Collections.NonGeneric" version="4.0.1" targetFramework="monoandroid80" />
-  <package id="System.Console" version="4.3.0" targetFramework="monoandroid80" />
-  <package id="System.Dynamic.Runtime" version="4.0.11" targetFramework="monoandroid80" />
-  <package id="System.Globalization" version="4.0.11" targetFramework="monoandroid80" />
-  <package id="System.Linq" version="4.1.0" targetFramework="monoandroid80" />
-  <package id="System.Reflection" version="4.1.0" targetFramework="monoandroid80" />
-  <package id="System.Reflection.Extensions" version="4.0.1" targetFramework="monoandroid80" />
-  <package id="System.Runtime" version="4.1.0" targetFramework="monoandroid80" />
-  <package id="System.Runtime.Extensions" version="4.1.0" targetFramework="monoandroid80" />
-  <package id="System.Text.RegularExpressions" version="4.1.0" targetFramework="monoandroid80" />
-  <package id="System.Threading" version="4.0.11" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Animated.Vector.Drawable" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Annotations" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Compat" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Core.UI" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Core.Utils" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Fragment" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Media.Compat" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.v7.AppCompat" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Android.Support.Vector.Drawable" version="25.3.1" targetFramework="monoandroid80" />
-  <package id="Xamarin.Build.Download" version="0.4.3" targetFramework="monoandroid80" />
-</packages>
