--- TestProjects/Logging.Core/Logging.iOS/Assets.xcassets/Contents.json
@@ -1,6 +1,0 @@
-﻿{
-  "info" : {
-    "version" : 1,
-    "author" : "xcode"
-  }
-}
