--- TestProjects/Logging.Core/Logging.iOS/ViewController.cs
@@ -1,26 +1,0 @@
-﻿using System;
-
-using UIKit;
-
-namespace Logging.iOS
-{
-    public partial class ViewController : UIViewController
-    {
-        protected ViewController(IntPtr handle) : base(handle)
-        {
-            // Note: this .ctor should not contain any initialization logic.
-        }
-
-        public override void ViewDidLoad()
-        {
-            base.ViewDidLoad();
-            // Perform any additional setup after loading the view, typically from a nib.
-        }
-
-        public override void DidReceiveMemoryWarning()
-        {
-            base.DidReceiveMemoryWarning();
-            // Release any cached data, images, etc that aren't in use.
-        }
-    }
-}
