--- TestProjects/Logging.Core/Logging.iOS/packages.config
@@ -1,26 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Microsoft.CSharp" version="4.0.1" targetFramework="xamarinios10" />
-  <package id="MvvmCross" version="5.3.2" targetFramework="xamarinios10" />
-  <package id="MvvmCross.Binding" version="5.3.2" targetFramework="xamarinios10" />
-  <package id="MvvmCross.Core" version="5.3.2" targetFramework="xamarinios10" />
-  <package id="MvvmCross.Platform" version="5.3.2" targetFramework="xamarinios10" />
-  <package id="MvvmCross.StarterPack" version="5.3.2" targetFramework="xamarinios10" />
-  <package id="Serilog" version="2.5.0" targetFramework="xamarinios10" />
-  <package id="Serilog.Sinks.Console" version="3.0.1" targetFramework="xamarinios10" />
-  <package id="Serilog.Sinks.Literate" version="3.0.1-dev-00044" targetFramework="xamarinios10" />
-  <package id="Serilog.Sinks.Xamarin" version="0.1.29" targetFramework="xamarinios10" />
-  <package id="System.Collections" version="4.0.11" targetFramework="xamarinios10" />
-  <package id="System.Collections.NonGeneric" version="4.0.1" targetFramework="xamarinios10" />
-  <package id="System.Console" version="4.3.0" targetFramework="xamarinios10" />
-  <package id="System.Dynamic.Runtime" version="4.0.11" targetFramework="xamarinios10" />
-  <package id="System.Globalization" version="4.0.11" targetFramework="xamarinios10" />
-  <package id="System.Linq" version="4.1.0" targetFramework="xamarinios10" />
-  <package id="System.Reflection" version="4.1.0" targetFramework="xamarinios10" />
-  <package id="System.Reflection.Extensions" version="4.0.1" targetFramework="xamarinios10" />
-  <package id="System.Runtime" version="4.1.0" targetFramework="xamarinios10" />
-  <package id="System.Runtime.Extensions" version="4.1.0" targetFramework="xamarinios10" />
-  <package id="System.Text.RegularExpressions" version="4.1.0" targetFramework="xamarinios10" />
-  <package id="System.Threading" version="4.0.11" targetFramework="xamarinios10" />
-  <package id="Xamarin.TestCloud.Agent" version="0.20.3" targetFramework="xamarinios10" />
-</packages>
