+++ TestProjects/Playground/Playground.Droid/Setup.cs
@@ -3,7 +3,9 @@ using Android.Content;
 using MvvmCross.Core.ViewModels;
 using MvvmCross.Droid.Support.V7.AppCompat;
 using MvvmCross.Droid.Views;
+using MvvmCross.Platform.Logging;
 using Playground.Core;
+using Serilog;
 
 namespace Playground.Droid
 {
@@ -21,6 +23,18 @@ namespace Playground.Droid
         protected override IMvxAndroidViewPresenter CreateViewPresenter()
         {
             return new MvxAppCompatViewPresenter(AndroidViewAssemblies);
+        }
+
+        protected override MvxLogProviderType GetDefaultLogProviderType()
+            => MvxLogProviderType.Serilog;
+
+        protected override IMvxLogProvider CreateLogProvider()
+        {
+            Log.Logger = new LoggerConfiguration()
+                .MinimumLevel.Debug()
+                .WriteTo.AndroidLog()
+                .CreateLogger();
+            return base.CreateLogProvider();
         }
     }
 }
