+++ TestProjects/Playground/Playground.Droid/packages.config
@@ -1,5 +1,19 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
+  <package id="Microsoft.CSharp" version="4.0.1" targetFramework="monoandroid71" />
+  <package id="Serilog" version="2.5.0" targetFramework="monoandroid71" />
+  <package id="Serilog.Sinks.Xamarin" version="0.1.29" targetFramework="monoandroid71" />
+  <package id="System.Collections" version="4.0.11" targetFramework="monoandroid71" />
+  <package id="System.Collections.NonGeneric" version="4.0.1" targetFramework="monoandroid71" />
+  <package id="System.Dynamic.Runtime" version="4.0.11" targetFramework="monoandroid71" />
+  <package id="System.Globalization" version="4.0.11" targetFramework="monoandroid71" />
+  <package id="System.Linq" version="4.1.0" targetFramework="monoandroid71" />
+  <package id="System.Reflection" version="4.1.0" targetFramework="monoandroid71" />
+  <package id="System.Reflection.Extensions" version="4.0.1" targetFramework="monoandroid71" />
+  <package id="System.Runtime" version="4.1.0" targetFramework="monoandroid71" />
+  <package id="System.Runtime.Extensions" version="4.1.0" targetFramework="monoandroid71" />
+  <package id="System.Text.RegularExpressions" version="4.1.0" targetFramework="monoandroid71" />
+  <package id="System.Threading" version="4.0.11" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Animated.Vector.Drawable" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Annotations" version="25.3.1" targetFramework="monoandroid71" />
   <package id="Xamarin.Android.Support.Compat" version="25.3.1" targetFramework="monoandroid71" />
