+++ TestProjects/Playground/Playground.Forms.Droid/Setup.cs
@@ -6,6 +6,8 @@ using MvvmCross.Droid.Views;
 using MvvmCross.Forms.Droid;
 using MvvmCross.Forms.Droid.Platform;
 using MvvmCross.Forms.Platform;
+using MvvmCross.Platform.Logging;
+using Serilog;
 
 namespace Playground.Forms.Droid
 {
@@ -13,6 +15,18 @@ namespace Playground.Forms.Droid
     {
         public Setup(Context applicationContext) : base(applicationContext)
         {
+        }
+
+        protected override MvxLogProviderType GetDefaultLogProviderType()
+            => MvxLogProviderType.Serilog;
+
+        protected override IMvxLogProvider CreateLogProvider()
+        {
+            Log.Logger = new LoggerConfiguration()
+                .MinimumLevel.Debug()
+                .WriteTo.AndroidLog()
+                .CreateLogger();
+            return base.CreateLogProvider();
         }
 
         protected override MvxFormsApplication CreateFormsApplication()
